requirements.txt
lxml
psutil
PyMySQL
PyYAML
requests
selenium
webdriver_manager
xmltodict
packaging
joblib
uuid
